+++ NEWorld.vcxproj
@@ -174,6 +174,7 @@
     <ClCompile Include="include\Netycat\src\Netycat\Core\Socket.cpp" />
     <ClCompile Include="source\Chunk.cpp" />
     <ClCompile Include="source\ChunkPtrArray.cpp" />
+    <ClCompile Include="source\CreateWorldMenu.cpp" />
     <ClCompile Include="source\GUIOptions.cpp" />
     <ClCompile Include="source\HeightMap.cpp" />
     <ClCompile Include="source\Definitions.cpp" />
