+++ NEWorld.vcxproj.filters
@@ -217,5 +217,8 @@
     <ClCompile Include="source\GUIOptions.cpp">
       <Filter>源代码\GUI\Menus</Filter>
     </ClCompile>
+    <ClCompile Include="source\CreateWorldMenu.cpp">
+      <Filter>源代码\GUI\Menus</Filter>
+    </ClCompile>
   </ItemGroup>
 </Project>
