+++ source/Menus.h
@@ -8,17 +8,8 @@
 #include <shellapi.h>
 
 template<typename T>
-string strWithVar(string str, T var){
-	std::stringstream ss;
-	ss << str << var;
-	return ss.str();
-}
-int getDotCount(string s) {
-	int ret = 0;
-	for (unsigned int i = 0; i != s.size(); i++)
-	if (s[i] == '.') ret++;
-	return ret;
-}
+string strWithVar(string str, T var);
+int getDotCount(string s);
 
 
 void mainmenu();
@@ -27,6 +18,7 @@ void Renderoptions();
 void GUIoptions();
 void worldmenu();
 void createworldmenu();
+void MultiplayerGameMenu();
 
 class NEMultiplayerGameMenu : public gui::UIView
 {
@@ -50,6 +42,11 @@ public:
 class MainMenu : public gui::UIView
 {
 private:
+	int leftp = windowwidth / 2 - 200;
+	int midp = windowwidth / 2;
+	int rightp = windowwidth / 2 + 200;
+	int upp = 280;
+
 	gui::UIButton startbtn = gui::UIButton("单人游戏");
 	gui::UIButton multiplayerbtn = gui::UIButton("多人游戏");
 	gui::UIButton optionsbtn = gui::UIButton(">> 选项...");
@@ -70,11 +67,9 @@ private:
 	gui::UITrackBar  FOVyBar = gui::UITrackBar(strWithVar("视野角度：", FOVyNormal), 120, (int)(FOVyNormal - 1));
 	gui::UITrackBar  mmsBar  = gui::UITrackBar(strWithVar("鼠标灵敏度：", mousemove), 120, (int)(mousemove * 40 * 2 - 1));
 	gui::UITrackBar  viewdistBar = gui::UITrackBar(strWithVar("渲染距离：", viewdistance), 120, (viewdistance - 1) * 8 - 1);
-	//gui::UIButton*	ciArrayBtn = UIButton("使用区块索引数组：" + boolstr(UseCIArray))
-	gui::UIButton	rdstbtn  = gui::UIButton(">> 渲染选项...");
-	gui::UIButton	gistbtn  = gui::UIButton(">> 图形界面选项...");
-	gui::UIButton	backbtn  = gui::UIButton("<< 返回主菜单");
-	//gui::UIButton*	savebtn = UIButton("保存设置")
+	gui::UIButton	 rdstbtn  = gui::UIButton(">> 渲染选项...");
+	gui::UIButton	 gistbtn  = gui::UIButton(">> 图形界面选项...");
+	gui::UIButton	 backbtn  = gui::UIButton("<< 返回主菜单");
 public:
 	Options();
 	~Options();
@@ -115,12 +110,48 @@ public:
 	void OnRender();
 };
 
+class NEWorldMenu : public gui::UIView
+{
+private:
+	bool refresh = true;
+	int selected = 0, mouseon;
+	ubyte mblast = 1;
+	int  mwlast = 0;
+	string chosenWorldName;
+	vector<string> worldnames;
+	vector<TextureID> thumbnails, texSizeX, texSizeY;
+	int trs = 0;
+	int worldcount;
+	int leftp = windowwidth / 2 - 200;
+	int midp = windowwidth / 2;
+	int rightp = windowwidth / 2 + 200;
+	int downp = windowheight - 20;
+
+	gui::UILabel          title = gui::UILabel("==============<  选 择 世 界  >==============");
+	gui::UIVerticalScroll UIVerticalScroll = gui::UIVerticalScroll(100, 0);
+	gui::UIButton         enterbtn = gui::UIButton("进入选定的世界");
+	gui::UIButton         deletebtn = gui::UIButton("删除选定的世界");
+	gui::UIButton         backbtn = gui::UIButton("<< 返回主菜单");
+
+public:
+	NEWorldMenu();
+	~NEWorldMenu();
+
+	void OnResize();
+	void OnUpdate();
+	void OnRender();
+};
+
 class CreateWorldMenu : public gui::UIView
 {
 private:
-	gui::UILabel  title = gui::UILabel("===============< 图形界面选项 >==============");
-	gui::UIButton fontbtn = gui::UIButton("全部使用Unicode字体：" + boolstr(TextRenderer::useUnicodeASCIIFont));
-	gui::UIButton backbtn = gui::UIButton("<< 返回选项菜单");
+	gui::UILabel   title       = gui::UILabel("==============<  新 建 世 界  >==============");
+	gui::UITextBox worldnametb = gui::UITextBox("输入世界名称");
+	gui::UIButton  okbtn       = gui::UIButton("确定");
+	gui::UIButton  backbtn = gui::UIButton("<< 返回世界菜单");
+
+	bool worldnametbChanged = false;
+
 
 public:
 	CreateWorldMenu();
@@ -130,5 +161,6 @@ public:
 	void OnUpdate();
 	void OnRender();
 };
+
